Rename `es2017.js` to `es2020.js`
